+++ quint/src/runtime/impl/compilerImpl.ts
@@ -87,7 +87,7 @@ export class CompilerVisitor implements IRVisitor {
   private execListener: ExecutionListener
 
   constructor(lookupTable: LookupTable, types: Map<bigint, TypeScheme>,
-      rand: () => number, listener: ExecutionListener) {
+      rand: (bound: bigint) => bigint, listener: ExecutionListener) {
     this.lookupTable = lookupTable
     this.types = types
     this.rand = rand
@@ -640,7 +640,7 @@ export class CompilerVisitor implements IRVisitor {
       case 'size':
         this.applyFun(app.id,
           1,
-          set => just(rv.mkInt(BigInt(set.cardinality()))))
+          set => set.cardinality().map(rv.mkInt))
         break
 
       case 'isFinite':
@@ -1293,19 +1293,24 @@ export class CompilerVisitor implements IRVisitor {
       })
 
       const ncandidates = successors.length
+      let choice
       if (ncandidates === 0) {
         // no successor: restore the state and return false
         this.recoverNextVars(valuesBefore)
         this.execListener.onAnyReturn(args.length, -1)
         return just(rv.mkBool(false))
+      } else if (ncandidates === 1) {
+        // There is exactly one successor, the execution is deterministic.
+        // No need for randomization. This may reduce the number of tests.
+        choice = 0
       } else {
         // randomly pick a successor and return true
-        // https://stackoverflow.com/questions/4959975/generate-random-number-between-two-numbers-in-javascript
-        const choice = Math.floor(this.rand() * ncandidates)
-        this.recoverNextVars(successors[choice])
-        this.execListener.onAnyReturn(args.length, successorIndices[choice])
-        return just(rv.mkBool(true))
+        choice = Number(this.rand(BigInt(ncandidates)))
       }
+
+      this.recoverNextVars(successors[choice])
+      this.execListener.onAnyReturn(args.length, successorIndices[choice])
+      return just(rv.mkBool(true))
     }
 
     this.compStack.push(mkFunComputable(lazyCompute))
@@ -1316,13 +1321,16 @@ export class CompilerVisitor implements IRVisitor {
     this.applyFun(sourceId,
       1,
       set => {
-        const elem = set.pick(this.rand())
-        if (elem) {
-          return just(elem)
-        } else {
-          this.addRuntimeError(sourceId, `Applied oneOf to an empty set`)
-          return none()
+        const sizeOrNone = set.cardinality()
+        if (sizeOrNone.isJust()) {
+          if (sizeOrNone.value === 0n) {
+            this.addRuntimeError(sourceId, `Applied oneOf to an empty set`)
+            return none()
+          }
+          return set.pick(this.rand(sizeOrNone.value))
         }
+        // an infinite set, pick an integer from the range [-2^255, 2^255)
+        return set.pick(-(2n ** 255n) + this.rand(2n ** 256n))
       }
     )
   }